refactor(mmctl): extract engage-admin command to separate files
Move engage-admin command and tests out of the shared roles.go and
roles_test.go into dedicated files to minimize merge conflicts:

- roles_engage_admin.go: command definition and implementation
- roles_engage_admin_test.go: unit tests

Each file registers itself via its own init() function.
